Start moving instrumentation items out of server crate
